* use object based TTree instead of plain ntuples * add HTTDataFormats package to make the dictionaries * migrate MiniAODVertexAnalyzer to new format
