+++ HTT/plugins/SealModule.cc
@@ -1,0 +1,9 @@
+// Define plugin
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+#include "WarsawAnalysis/HTT/plugins/MiniAODVertexAnalyzer.h"
+#include "WarsawAnalysis/HTT/plugins/MiniAODGenTauTauAnalyzer.h"
+
+DEFINE_FWK_MODULE(MiniAODVertexAnalyzer);
+DEFINE_FWK_MODULE(MiniAODGenTauTauAnalyzer);
+
