+++ HTT/plugins/SealModule.cc~
@@ -1,0 +1,3 @@
+// Define plugin
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(MiniAODVertexAnalyzer);
