+++ HTT/test/prod/crab_genTuple.cfg
@@ -1,0 +1,37 @@
+[CMSSW]
+datasetpath             = 
+total_number_of_events  = -1
+events_per_job          = 500000
+#number_of_jobs          = 5
+#use_parent = 1
+
+pset                    = ../gen_ana_cfg.py
+
+#output_file             = hltOut.root,triggerAnalyzerResult.root
+output_file             = genTauTauAna.root
+
+[USER]
+return_data             = 0
+copy_data               = 1
+publish_data            = 0
+publish_data_name       = HTTCP_genTuple_v1
+#dbs_url_for_publication =  https://cmsdbsprod.cern.ch:8443/cms_dbs_ph_analysis_01_writer/servlet/DBSServlet
+storage_element         = T2_PL_Swierk
+user_remote_dir         = HTTCP_genTuple_v1
+ui_working_dir          = HTTCP_genTuple_v1
+#additional_input_files  = 
+
+[CRAB]
+jobtype    = cmssw
+scheduler  = remoteGlidein
+#scheduler  = glite
+use_server = 0
+
+[GRID]
+#ce_black_list=grid36.lal.in2p3.fr
+#se_white_list=node12.datagrid.cea.fr
+#dont_check_proxy= 1
+#user has to take care of the proxy.
+#virtual_organization= vo.llr.in2p3.fr
+#ce_white_list= srm-cms.cern.ch,ccsrm.in2p3.fr,caf.cern.ch
+#ce_black_list=heplnx206.pp.rl.ac.uk
