+++ README.md
@@ -1,2 +1,2 @@
-# usercode-WawTools
-git clone https://github.com/mbluj/usercode-WawTools.git WarsawAnalysis/HTT
+
+git clone https://github.com/mbluj/usercode-WawTools.git WarsawAnalysis
