--- test/prod/multicrab_genTuple.cfg
@@ -1,53 +1,0 @@
-[MULTICRAB]
-cfg=crab_genTuple.cfg
-
-[COMMON]
-USER.storage_element = T2_PL_Swierk
-# different PSet with different GT
-CMSSW.pset           = ../gen_ana_cfg.py
-# use user data from phys03 instance of DBS (uncomment here or add to specific items)
-#CMSSW.dbs_url        = phys03
-# uncomment to use parent dataset 
-#CMSSW.use_parent     = 1
-#comment out following to store output at SE
-#USER.return_data     = 1
-#USER.copy_data       = 0
-#change the following to publish data
-USER.publish_data    = 0
-#USER.publish_data    = 1
-# comment out the following to get edm-output
-CMSSW.get_edm_output = 0
-CMSSW.output_file    = genTauTauAna.root
-
-[DYToTauTau_v1]
-CMSSW.dbs_url     = phys03
-CMSSW.datasetpath = /DYToTauTau_M-50_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v1/bluj-DYToTauTau_M-50_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v1-76d179419fbff7169973d9fdfd1ce825/USER
-CMSSW.total_number_of_events = -1
-CMSSW.events_per_job = 200000
-USER.publish_data_name = HTTCP_genTuple_v1 
-USER.user_remote_dir = HTTCP_genTuple_DY_v1
-
-[GGH125ToTauTau_v1]
-CMSSW.dbs_url     = phys03
-CMSSW.datasetpath = /SUSYGluGluH0ToTauTau_M-125_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v2/bluj-SUSYGluGluH0ToTauTau_M-125_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v2-17aeb1344d1bc358be14f25dc5a8c422/USER
-CMSSW.total_number_of_events = -1
-CMSSW.events_per_job = 200000
-USER.publish_data_name = HTTCP_genTuple_v1 
-USER.user_remote_dir = HTTCP_genTuple_GGH125_v1
-
-[GGA125ToTauTau_v1]
-CMSSW.dbs_url     = phys03
-CMSSW.datasetpath = /SUSYGluGluA0ToTauTau_M-125_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v1/bluj-SUSYGluGluA0ToTauTau_M-125_TuneCUETP8M1_13TeV_pythia8_GEN_RunIIWinter15GS-MCRUN2_71_V1-v1-ca5c4cbeea1a7b632759b4a8c2b92091/USER
-CMSSW.total_number_of_events = -1
-CMSSW.events_per_job = 200000
-USER.publish_data_name = HTTCP_genTuple_v1
-USER.user_remote_dir = HTTCP_genTuple_GGA125_v1
-
-[GGA120ToTauTau_Mini_v1]
-CMSSW.dbs_url     = https://cmsweb.cern.ch/dbs/prod/global/DBSReader
-CMSSW.pset        = ../gen_ana_mini_cfg.py
-CMSSW.datasetpath = /SUSYGluGluToHToTauTau_M-120_TuneCUETP8M1_13TeV-pythia8/RunIISpring15DR74-Asympt25ns_MCRUN2_74_V9-v1/MINIAODSIM
-CMSSW.total_number_of_events = -1
-CMSSW.events_per_job = 200000
-USER.publish_data_name = HTTCP_genTuple_v1
-USER.user_remote_dir = HTTCP_genTuple_GGA120_Mini_v1
